move C-preprocessor shared code into infrastructure.py.
